Simple filter of RGB image. CV mat should move memory not clone it
I created a fast CL pipeline to treat images by first adding an alpha
channel to the RGB image in order for it to be read in the kernel.
We created alligned memory for the buffers in order for them to be
mapped and unmapped with zero time.
Next stp would be that the CV mat_out should do a move operation on the
buffer and not just deep clone everything because that is slow.